+++ internal/framework/services/objectstorage/resource_bucket_public_access_block.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/objectstorage"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
@@ -80,17 +81,17 @@ func (r *bucketPublicAccessBlockResource) Configure(_ context.Context, req resou
 		return
 	}
 
-	client, ok := req.ProviderData.(*objectstorage.Client)
+	clientBundle, ok := req.ProviderData.(*services.SdkBundle)
 	if !ok {
 		resp.Diagnostics.AddError(
 			"Unexpected Data Source Configure Type",
-			fmt.Sprintf("Expected *objectstorage.Client, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+			fmt.Sprintf("Expected *services.SdkBundle, got: %T. Please report this issue to the provider developers.", req.ProviderData),
 		)
 
 		return
 	}
 
-	r.client = client
+	r.client = clientBundle.S3Client
 }
 
 // Create creates the bucket.